refactor: centralize UI config and expose input controller
